requirements.txt
absl-py==2.0.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
cmake==3.27.6
einops==0.7.0
ema-pytorch==0.2.3
filelock==3.12.4
ftfy==6.1.1
google-auth==2.23.2
google-auth-oauthlib==1.0.0
grpcio==1.59.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
lit==17.0.1
lmdb==1.4.1
Markdown==3.4.4
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.1.78
Pillow==10.0.1
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy==1.10.1
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
urllib3==2.0.5
wcwidth==0.2.8
Werkzeug==3.0.0
zipp==3.17.0
